% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ref_id_table}
\alias{ref_id_table}
\title{Reference ID table of all included reviews}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 3 columns.
}
\usage{
ref_id_table
}
\description{
Reference ID table of all included reviews
}
\keyword{datasets}
